fix bugs

+++ Dtos/AllTrainingSessionDto.cs
@@ -7,5 +7,5 @@ public class AllTrainingSessionDto
     public int Id { get; set; }
     public int DayNumber { get; set; }
     public string TrainingSessionStatus { get; set; } 
-    public required byte[] ExerciseCompletionBitmap { get; set; } 
+    public required int[] ExerciseCompletionBitmap { get; set; } 
 }
