+++ Dtos/TrainingSessionDto.cs
@@ -11,7 +11,7 @@ namespace sport_app_backend.Dtos
         public required TrainingSessionProgramInDayDto ProgramInDay { get; set; }
         public int DayNumber { get; set; }
         public string TrainingSessionStatus { get; set; } 
-        public required byte[] ExerciseCompletionBitmap { get; set; } 
+        public required int[] ExerciseCompletionBitmap { get; set; } 
     
     }
 }
